\name{sondowania-package}
\alias{sondowania-package}
\alias{sondowania}
\docType{package}
\title{
\packageTitle{sondowania}
}
\description{
\packageDescription{sondowania}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sondowania}
\packageIndices{sondowania}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{sondowania}

Maintainer: \packageMaintainer{sondowania}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
